% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation}
\alias{bifurcation}
\title{bifurcation}
\usage{
bifurcation(
  s0,
  p0,
  m,
  p_i = 1L,
  s_i = 1L,
  step = 0.01,
  p_range = c(0, 1),
  s_range = c(0, 1.1),
  log_scale_axis = "",
  positive = FALSE,
  steady_warnings = FALSE,
  ...
)
}
\description{
bifurcation
}
